refactor rendering-related files to DDD organization
