+++ src/composables/canvas/useCanvasTransformSync.ts
@@ -1,7 +1,7 @@
 import { onUnmounted, ref } from 'vue'
 
 import { LGraphCanvas } from '@/lib/litegraph/src/litegraph'
-import { useCanvasStore } from '@/stores/graphStore'
+import { useCanvasStore } from '@/renderer/core/canvas/canvasStore'
 
 interface CanvasTransformSyncOptions {
   /**
