+++ src/renderer/core/canvas/canvasStore.ts
